skills obj data file with imports moved, functioning
